+++ R/explore.R
@@ -454,7 +454,10 @@ explore <- function(
                                     bio = bio)
 
     if (any(sapply(valid.detections, function(x) any(!is.na(x$Sensor.Value))))) {
-      sensor.plots <- printSensorData(detections = valid.detections, rsp.info = rsp.info)
+      sensor.plots <- printSensorData(detections = valid.detections, 
+                                      spatial = spatial,
+                                      rsp.info = rsp.info, 
+                                      type = plot.detections.by)
     } else {
       sensor.plots <- NULL
     }